requirements.txt
bcrypt==3.2.0
cachelib==0.6.0
cffi==1.15.0
click==8.0.4
Flask==2.0.3
Flask-Cors==3.0.10
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
psycopg2==2.9.3
pycparser==2.21
python-dotenv==0.19.2
six==1.16.0
SQLAlchemy==1.4.31
Werkzeug==2.0.3
